+++ src/base-layers/rectangle/rectangle-layer.ts
@@ -73,13 +73,13 @@ export class RectangleLayer<
           anchorEffect[1] = rectangle.anchor.y || 0;
         }
         const topLeft = [
-          rectangle.x - anchorEffect[0],
-          rectangle.y - anchorEffect[1]
+          rectangle.position[0] - anchorEffect[0],
+          rectangle.position[1] - anchorEffect[1]
         ];
 
         return new Bounds({
-          height: rectangle.height,
-          width: rectangle.width,
+          height: rectangle.size[1],
+          width: rectangle.size[0],
           x: topLeft[0],
           y: topLeft[1]
         });
@@ -116,14 +116,14 @@ export class RectangleLayer<
             }
 
             const topLeft = [
-              rectangle.x - anchorEffect[0] / maxScale,
-              rectangle.y - anchorEffect[1] / maxScale
+              rectangle.position[0] - anchorEffect[0] / maxScale,
+              rectangle.position[1] - anchorEffect[1] / maxScale
             ];
 
             // Reverse project the size and we should be within the distorted world coordinates
             return new Bounds({
-              height: rectangle.height / maxScale,
-              width: rectangle.width / maxScale,
+              height: rectangle.size[1] / maxScale,
+              width: rectangle.size[0] / maxScale,
               x: topLeft[0],
               y: topLeft[1]
             }).containsPoint(point);
@@ -140,7 +140,7 @@ export class RectangleLayer<
           }
 
           const topLeft = subtract2(
-            [rectangle.x, rectangle.y],
+            [rectangle.position[0], rectangle.position[1]],
             divide2(anchorEffect, projection.camera.scale)
           );
 
@@ -148,8 +148,8 @@ export class RectangleLayer<
 
           // Reverse project the size and we should be within the distorted world coordinates
           return new Bounds({
-            height: rectangle.height,
-            width: rectangle.width,
+            height: rectangle.size[1],
+            width: rectangle.size[0],
             x: topLeft[0],
             y: topLeft[1]
           }).containsPoint(screenPoint);
@@ -192,7 +192,7 @@ export class RectangleLayer<
           easing: animate.location,
           name: RectangleLayer.attributeNames.location,
           size: InstanceAttributeSize.TWO,
-          update: o => [o.x, o.y]
+          update: o => o.position
         },
         {
           name: RectangleLayer.attributeNames.anchor,
@@ -202,7 +202,7 @@ export class RectangleLayer<
         {
           name: RectangleLayer.attributeNames.size,
           size: InstanceAttributeSize.TWO,
-          update: o => [o.width, o.height]
+          update: o => o.size
         },
         {
           name: RectangleLayer.attributeNames.depth,