+++ src/util/index.ts
@@ -1,5 +1,4 @@
 export * from "./common-options";
-export * from "./animation-helper";
 export * from "./auto-easing-method";
 export * from "./chart-camera";
 export * from "./extend-shader-io";
@@ -13,3 +12,4 @@ export * from "./matrix";
 export * from "./view-camera";
 export * from "./next-frame";
 export * from "./easing-util";
+export * from "./promise-resolver";