+++ test/kitchen-sink/examples/bounded-view3.ts
@@ -18,7 +18,7 @@ import { BaseExample, TestResourceKeys } from "./base-example";
 export class BoundedView3 extends BaseExample {
   view: string;
   manager: BasicCameraController;
-  originalRange: Bounds;
+  originalRange: Bounds<never>;
 
   makeController(
     defaultCamera: ChartCamera,
@@ -91,12 +91,10 @@ export class BoundedView3 extends BaseExample {
           type: anchor[i]
         },
         color: [color[i][0], color[i][1], color[i][2], color[i][3]],
-        height: height[i],
         id: `rectangle${i}`,
         scaling: ScaleMode.ALWAYS,
-        width: width[i],
-        x: x[i],
-        y: y[i]
+        size: [width[i], height[i]],
+        position: [x[i], y[i]]
       });
 
       rectangleProvider.add(rectangle);